requirements.txt
websockets==11.0.3
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pymysql==1.1.0
sqlalchemy==2.0.23
colorlog==6.7.0 
cryptography
aiohttp==3.9.1